Delete message functionality working!!!
Original problem: The only delete button that was working was the one for the message at the top of the list. I needed them to work for any message in any position.

Solution: I created each element separately using document.createElement(). After creating all elements and divs, I nested them in the desired order and then appended the completed message div to the messages list. Originally, I was putting all of these elements into one big innerHTML statement and trying to add event listeners to the buttons was difficult. I found that I already had this functionality working in the basic version (double click to delete), so I tried to replicate what I had already done, but on a larger scale to solve this problem.