Delete button functionality update
Delete button should now only appear on messages that have been published from that browser. Also fixes an issue that caused the delete button not to show up after a message was created, but would show up on future refreshes. This was done by adding the delete button creation code to the publish button's event listener after the myAuthoredMessages array had time to process the ID of the new message.